Registering VirtualNetworksStore in constructor instead of StoreConfig
