+++ src/js/constants/StoreConfig.js
@@ -2,13 +2,9 @@ import {
   MESOS_STATE_CHANGE,
   MESOS_STATE_REQUEST_ERROR,
 
-  VIRTUAL_NETWORKS_CHANGE,
-  VIRTUAL_NETWORKS_REQUEST_ERROR,
-
   VISIBILITY_CHANGE
 } from './EventTypes';
 import MesosStateStore from '../stores/MesosStateStore';
-import VirtualNetworksStore from '../stores/VirtualNetworksStore';
 import VisibilityStore from '../stores/VisibilityStore';
 
 const ListenersDescription = {
@@ -22,18 +18,6 @@ const ListenersDescription = {
       if (event === 'success') {
         return Object.keys(store.get('lastMesosState')).length;
       }
-    },
-    listenAlways: true
-  },
-
-  virtualNetworks: {
-    store: VirtualNetworksStore,
-    events: {
-      success: VIRTUAL_NETWORKS_CHANGE,
-      error: VIRTUAL_NETWORKS_REQUEST_ERROR
-    },
-    unmountWhen: function () {
-      return true;
     },
     listenAlways: true
   },
